First pass at editing unused styles in the wizard
* Adds a styleid variable to the wizard widgets, so it is not always
  assumming the current style.
* Gives an indication that the style being editing in the wizard is not the
  current style

TODO: Needs extensive testing, where every page/setting type is tested for
the current style and a not-in-use style to make sure when saved it goes
to the right place.

TODO: The javascript for the inline style layout selector widget does
not work with this code and will need to be either fixed or disabled
for working on unused styles.